Calculate price for strategy3

+++ src/main/java/application/dak/DAK/backend/common/models/StrategyPrice.java
@@ -1,5 +1,5 @@
 package application.dak.DAK.backend.common.models;
 
 public interface StrategyPrice {
-    Float execute();
+    double execute();
 }
