+++ src/main/java/application/dak/DAK/backend/common/models/StrategyPrice2.java
@@ -9,7 +9,7 @@ public class StrategyPrice2 implements StrategyPrice {
     public Integer tripTax;
 
     @Override
-    public Float execute() {
+    public double execute() {
         return (float) (tripTax * KG);
     }
 }
